refactor: convert layout creation tests back to being unit tests
