+++ tests/integration/main.rs
@@ -4,7 +4,6 @@ mod util;
 
 mod arg_tests;
 mod invalid_config_tests;
-mod layout_management_tests;
 mod layout_movement_tests;
 
 #[cfg(all(target_arch = "x86_64", target_os = "linux"))]
